Added Cronjob For updating the Fund
