+++ campaigns/views.py
@@ -120,14 +120,3 @@ class CampaignDetailsApi(GenericMethodsMixin,APIView):
 
 
 
-
-def update_campaign_fund():
-   campaigns = Campaign.objects.filter(status=CampaignChoices.PENDING)
-   for campaign in campaigns:
-        print(campaign)
-        total_donation = Donor.objects.filter(campaign=campaign).aggregate(total_amount=models.Sum('amount'))['total_amount']
-        campaign.fund_raised = total_donation if total_donation else 0
-        # campaign.save()
-        
-
-update_campaign_fund()
