+++ reuirements.txt
@@ -4,3 +4,5 @@ djangorestframework
 djangorestframework-simplejwt
 python-dotenv
 Pillow
+python
+django-crontab
